requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
scikit-fuzzy==0.4.2
pyyaml==6.0
tqdm==4.65.0
jupyter==1.0.0
imbalanced-learn==0.10.1
scikit-plot==0.3.7